+++ xclim/testing/tests/test_sdba/test_adjustment.py
@@ -533,7 +533,7 @@ class TestPrincipalComponents:
         assert (ref - scen).mean() < 5e-3
 
 
-@pytest.mark.slow
+@pytest.mark.xfail(reason="Inaccurate version.")
 class TestExtremeValues:
     @pytest.mark.parametrize(
         "c_thresh,q_thresh,frac,power,diags",