+++ HISTORY.rst
@@ -5,6 +5,10 @@ History
 0.29.0 (unreleased)
 -------------------
 
+Announcements
+~~~~~~~~~~~~~
+* It was found that the `ExtremeValues` adjustment algorithm was not as accurate and stable as first thought. It was hidden from `xclim.sdba` but it is still accessible through `xclim.sdba.adjustment`, with a warning. Work on improving the algorithm is on-going and a better version will be implemented in a future version.
+
 New indicators
 ~~~~~~~~~~~~~~
 * ``effective_growing_degree_days`` indice returns growing degree days using dynamic start and end dates for the growing season (based on Bootsma et al. (2005)). This has also been wrapped as an indicator.